libcusmm: jit kernels using a single omp thread
JIT the matrix-matrix multiplication kernels using a single thread.
Taking advantage of the fact that the table of handles to JITted kernels
are shared among openmp threads on the node, have only a single thread
perform the JITing.

Prior to this commit, redundant JITing of matrix-matrix multiplication
kernels was performed.

This fixes a performance issue in CP2K:
https://github.com/cp2k/cp2k/issues/155