Issue #214: added test case for DBCSR/ACC interface.
This program tests the ACC interface (include/acc.h) for adhering to expectations. The expected behavior is to match the CUDA based implementation, which was available first. This test program can serve as a specification for other backends such as the OpenMP based backend. It may also be used to stress-test any backend including the CUDA based backend for thread-safety. Thread-safety is an implicit requirement induced by DBCSR (and CP2K). To test any backend (other than the OpenMP based backend), the Makefile must be adjusted to link with the desired backend.